requirements.txt
asgiref==3.5.2
Django==4.1.4
django-braces==1.15.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
